merge OIModuleMaps and master (#8) (#9)
* Revert "Merge pull request #7 from kroner/OIModuleMaps"

This reverts commit 3dcb753d1e8e47f7b7a18b6ae3a6025ad8560af5, reversing
changes made to 2603b5bcc68c6b431a350e44a94d2738eaf74a52.

* Finished revising the package RegularPackages. I recommed changing the way Automatons are constructed. Particularly,
not making the final step of assigning all larroes that haven't been defined to a "Rejected" state.

When doing concatenations of unions these added states tend to make the automatons grow excessively. It would be
better to be able to make the concatenations and unions without this states. Then once the automaton is finished put
any undefined arrows to a new rejected state.